services/agent/requirements.txt
cryptography==42.0.8
requests==2.32.2
psycopg2-binary==2.9.9
pyyaml==6.0.2
tenacity==8.3.0
